Launch server when running suite
This is a quick and dirty solution to have a server running in order to
bootstrap a test harness.

It has a few ugly defaults:

This is ugly and not very failure-proof. We would need to refactor some logic
out of `bin/rest-ftp-daemon` executable to do this in a cleaner and more
seamless way.

We should be smarter when waiting for the server to be actually reachable.

We use port 5678 for now, but it might be already taken. Further care should be
taken to make sure we don't get into trouble because of this.

TODO:

* extract logic from clean up bin/rest-ftp-daemon
* avoid using `sleep` when launching server
* find available port instead of using hardcoded value 5678?